requirements.txt
ccxt
pandas
pandas_ta
python-telegram-bot